+++ TBG.Core/Interfaces/ITournamentEntry.cs
@@ -7,7 +7,7 @@ namespace TBG.Core.Interfaces
         int TournamentEntryId { get; set; }
         int TournamentId { get; set; }
         int TeamId { get; set; }
-        int Seed { get; set; }
+        double Seed { get; set; }
         ObservableCollection<IPerson> Members { get; set; }
     }
 }
