+++ TBG.Data/Tables/TournamentEntriesTable.cs
@@ -79,7 +79,7 @@ namespace TBG.Data.Tables
                 TournamentEntryId = Int32.Parse(reader["tournament_entry_id"].ToString()),
                 TournamentId = Int32.Parse(reader["tournament_id"].ToString()),
                 TeamId = Int32.Parse(reader["team_id"].ToString()),
-                Seed = Int32.Parse(reader["seed"].ToString())
+                Seed = Double.Parse(reader["seed"].ToString())
             };
         }
     }
