+++ TBG.UI/Classes/TournamentEntry.cs
@@ -13,7 +13,7 @@ namespace TBG.UI.Classes
         public int TournamentEntryId { get; set; }
         public int TournamentId { get; set; }
         public int TeamId { get; set; }
-        public int Seed { get; set; }
+        public double Seed { get; set; }
         public ObservableCollection<IPerson> Members { get; set; }
         //Need for XAML
         public string TeamName { get; set; }
